+++ src/main/java/pfr/center/models/DepartmentMapper.java
@@ -11,7 +11,7 @@ public class DepartmentMapper implements RowMapper<Department> {
     @Override
     public Department mapRow(ResultSet resultSet, int i) throws SQLException {
         Department district = new Department();
-        district.setId_depart(resultSet.getLong("ID_DEPART"));
+        district.setId_depart(resultSet.getInt("ID_DEPART"));
         district.setName_dep(resultSet.getString("NAME_DEP"));
         district.setNum_reg(resultSet.getString("NUM_REG"));
         district.setName_reg(resultSet.getString("NAME_REG"));
